requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
click==8.1.8
    # via code-annotations
code-annotations==2.2.0
    # via -r requirements/test.in
coverage[toml]==7.6.12
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-model-utils
django-model-utils==5.0.0
    # via -r requirements/base.txt
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-opaque-keys==2.11.0
    # via -r requirements/base.txt
factory-boy==3.3.3
    # via -r requirements/test.in
faker==36.1.1
    # via factory-boy
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via code-annotations
lxml[html-clean]==5.3.1
    # via
    #   -r requirements/base.txt
    #   lxml-html-clean
lxml-html-clean==0.4.1
    # via
    #   -r requirements/base.txt
    #   lxml
markupsafe==3.0.2
    # via jinja2
packaging==24.2
    # via pytest
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
pymongo==4.11.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.10.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
tzdata==2025.1
    # via faker

# The following packages are considered to be unsafe in a requirements file:
# setuptools
